[구름4.0] 소스 업데이트 - warning 코드 제거, jdk17 코드 개선 - 불필요 예외처리, 조건문 제거

+++ src/main/java/kr/gooroom/gpms/job/service/impl/JobDAO.java
@@ -43,7 +43,7 @@ public class JobDAO extends SqlSessionMetaDAO {
 	 */
 	public long createJobMaster(JobVO jobVO) throws SQLException {
 
-		return (long) sqlSessionMeta.insert("insertJobMaster", jobVO);
+		return sqlSessionMeta.insert("insertJobMaster", jobVO);
 
 	}
 
@@ -56,7 +56,7 @@ public class JobDAO extends SqlSessionMetaDAO {
 	 */
 	public long createJobTarget(JobVO jobVO) throws SQLException {
 
-		return (long) sqlSessionMeta.insert("insertJobTarget", jobVO);
+		return sqlSessionMeta.insert("insertJobTarget", jobVO);
 
 	}
